Moved two more kernels into separate files
