+++ platforms/common/src/CommonKernels.cpp
@@ -34,8 +34,6 @@
 #include "openmm/internal/ContextImpl.h"
 #include "openmm/internal/CustomCentroidBondForceImpl.h"
 #include "openmm/internal/CustomCompoundBondForceImpl.h"
-#include "openmm/internal/CustomHbondForceImpl.h"
-#include "openmm/internal/CustomManyParticleForceImpl.h"
 #include "openmm/internal/ThreadPool.h"
 #include "openmm/internal/timer.h"
 #include "CommonKernelSources.h"
@@ -2206,1052 +2204,6 @@ void CommonCalcGBSAOBCForceKernel::copyParametersToContext(ContextImpl& context,
     // Mark that the current reordering may be invalid.
     
     cc.invalidateMolecules(info, true, false);
-}
-
-class CommonCalcCustomHbondForceKernel::ForceInfo : public ComputeForceInfo {
-public:
-    ForceInfo(const CustomHbondForce& force) : force(force) {
-    }
-    bool areParticlesIdentical(int particle1, int particle2) {
-        return true;
-    }
-    int getNumParticleGroups() {
-        return force.getNumDonors()+force.getNumAcceptors()+force.getNumExclusions();
-    }
-    void getParticlesInGroup(int index, vector<int>& particles) {
-        int p1, p2, p3;
-        thread_local static vector<double> parameters;
-        if (index < force.getNumDonors()) {
-            force.getDonorParameters(index, p1, p2, p3, parameters);
-            particles.clear();
-            particles.push_back(p1);
-            if (p2 > -1)
-                particles.push_back(p2);
-            if (p3 > -1)
-                particles.push_back(p3);
-            return;
-        }
-        index -= force.getNumDonors();
-        if (index < force.getNumAcceptors()) {
-            force.getAcceptorParameters(index, p1, p2, p3, parameters);
-            particles.clear();
-            particles.push_back(p1);
-            if (p2 > -1)
-                particles.push_back(p2);
-            if (p3 > -1)
-                particles.push_back(p3);
-            return;
-        }
-        index -= force.getNumAcceptors();
-        int donor, acceptor;
-        force.getExclusionParticles(index, donor, acceptor);
-        particles.clear();
-        force.getDonorParameters(donor, p1, p2, p3, parameters);
-        particles.push_back(p1);
-        if (p2 > -1)
-            particles.push_back(p2);
-        if (p3 > -1)
-            particles.push_back(p3);
-        force.getAcceptorParameters(acceptor, p1, p2, p3, parameters);
-        particles.push_back(p1);
-        if (p2 > -1)
-            particles.push_back(p2);
-        if (p3 > -1)
-            particles.push_back(p3);
-    }
-    bool areGroupsIdentical(int group1, int group2) {
-        int p1, p2, p3;
-        thread_local static vector<double> params1, params2;
-        if (group1 < force.getNumDonors() && group2 < force.getNumDonors()) {
-            force.getDonorParameters(group1, p1, p2, p3, params1);
-            force.getDonorParameters(group2, p1, p2, p3, params2);
-            return (params1 == params2);
-        }
-        if (group1 < force.getNumDonors() || group2 < force.getNumDonors())
-            return false;
-        group1 -= force.getNumDonors();
-        group2 -= force.getNumDonors();
-        if (group1 < force.getNumAcceptors() && group2 < force.getNumAcceptors()) {
-            force.getAcceptorParameters(group1, p1, p2, p3, params1);
-            force.getAcceptorParameters(group2, p1, p2, p3, params2);
-            return (params1 == params2);
-        }
-        if (group1 < force.getNumAcceptors() || group2 < force.getNumAcceptors())
-            return false;
-        return true;
-    }
-private:
-    const CustomHbondForce& force;
-};
-
-CommonCalcCustomHbondForceKernel::~CommonCalcCustomHbondForceKernel() {
-    ContextSelector selector(cc);
-    if (donorParams != NULL)
-        delete donorParams;
-    if (acceptorParams != NULL)
-        delete acceptorParams;
-}
-
-static void applyDonorAndAcceptorForces(stringstream& apply, int atom, const string& value, bool trim=true) {
-    string forceNames[] = {"f1", "f2", "f3"};
-    string toAdd = (trim ? "trimTo3("+value+")" : value);
-    if (atom < 3)
-        apply << "localData[tbx+index]." << forceNames[atom]<<" += "<<toAdd<<";\n";
-    else
-        apply << forceNames[atom-3]<<" += "<<toAdd<<";\n";
-}
-
-void CommonCalcCustomHbondForceKernel::initialize(const System& system, const CustomHbondForce& force) {
-    // Record the lists of donors and acceptors, and the parameters for each one.
-
-    ContextSelector selector(cc);
-    int numContexts = cc.getNumContexts();
-    int startIndex = cc.getContextIndex()*force.getNumDonors()/numContexts;
-    int endIndex = (cc.getContextIndex()+1)*force.getNumDonors()/numContexts;
-    numDonors = endIndex-startIndex;
-    numAcceptors = force.getNumAcceptors();
-    if (numDonors == 0 || numAcceptors == 0)
-        return;
-    int numParticles = system.getNumParticles();
-    donors.initialize<mm_int4>(cc, numDonors, "customHbondDonors");
-    acceptors.initialize<mm_int4>(cc, numAcceptors, "customHbondAcceptors");
-    donorParams = new ComputeParameterSet(cc, force.getNumPerDonorParameters(), numDonors, "customHbondDonorParameters");
-    acceptorParams = new ComputeParameterSet(cc, force.getNumPerAcceptorParameters(), numAcceptors, "customHbondAcceptorParameters");
-    if (force.getNumGlobalParameters() > 0)
-        globals.initialize<float>(cc, force.getNumGlobalParameters(), "customHbondGlobals");
-    vector<vector<float> > donorParamVector(numDonors);
-    vector<mm_int4> donorVector(numDonors);
-    for (int i = 0; i < numDonors; i++) {
-        vector<double> parameters;
-        force.getDonorParameters(startIndex+i, donorVector[i].x, donorVector[i].y, donorVector[i].z, parameters);
-        donorParamVector[i].resize(parameters.size());
-        for (int j = 0; j < (int) parameters.size(); j++)
-            donorParamVector[i][j] = (float) parameters[j];
-    }
-    donors.upload(donorVector);
-    donorParams->setParameterValues(donorParamVector);
-    vector<vector<float> > acceptorParamVector(numAcceptors);
-    vector<mm_int4> acceptorVector(numAcceptors);
-    for (int i = 0; i < numAcceptors; i++) {
-        vector<double> parameters;
-        force.getAcceptorParameters(i, acceptorVector[i].x, acceptorVector[i].y, acceptorVector[i].z, parameters);
-        acceptorParamVector[i].resize(parameters.size());
-        for (int j = 0; j < (int) parameters.size(); j++)
-            acceptorParamVector[i][j] = (float) parameters[j];
-    }
-    acceptors.upload(acceptorVector);
-    acceptorParams->setParameterValues(acceptorParamVector);
-    info = new ForceInfo(force);
-    cc.addForce(info);
-
-    // Decide whether to use bounding boxes to accelerate the calculation.
-
-    int numDonorBlocks = (numDonors+31)/32;
-    int numAcceptorBlocks = (numAcceptors+31)/32;
-    useBoundingBoxes = (force.getNonbondedMethod() != CustomHbondForce::NoCutoff && numDonorBlocks*numAcceptorBlocks > cc.getNumThreadBlocks());
-    if (useBoundingBoxes) {
-        int elementSize = (cc.getUseDoublePrecision() ? sizeof(double) : sizeof(float));
-        donorBlockCenter.initialize(cc, numDonorBlocks, 4*elementSize, "donorBlockCenter");
-        donorBlockSize.initialize(cc, numDonorBlocks, 4*elementSize, "donorBlockSize");
-        acceptorBlockCenter.initialize(cc, numAcceptorBlocks, 4*elementSize, "acceptorBlockCenter");
-        acceptorBlockSize.initialize(cc, numAcceptorBlocks, 4*elementSize, "acceptorBlockSize");
-    }
-
-    // Record exclusions.
-
-    vector<mm_int4> donorExclusionVector(numDonors, mm_int4(-1, -1, -1, -1));
-    vector<mm_int4> acceptorExclusionVector(numAcceptors, mm_int4(-1, -1, -1, -1));
-    for (int i = 0; i < force.getNumExclusions(); i++) {
-        int donor, acceptor;
-        force.getExclusionParticles(i, donor, acceptor);
-        if (donor < startIndex || donor >= endIndex)
-            continue;
-        donor -= startIndex;
-        if (donorExclusionVector[donor].x == -1)
-            donorExclusionVector[donor].x = acceptor;
-        else if (donorExclusionVector[donor].y == -1)
-            donorExclusionVector[donor].y = acceptor;
-        else if (donorExclusionVector[donor].z == -1)
-            donorExclusionVector[donor].z = acceptor;
-        else if (donorExclusionVector[donor].w == -1)
-            donorExclusionVector[donor].w = acceptor;
-        else
-            throw OpenMMException("CustomHbondForce: this platform does not support more than four exclusions per donor");
-        if (acceptorExclusionVector[acceptor].x == -1)
-            acceptorExclusionVector[acceptor].x = donor;
-        else if (acceptorExclusionVector[acceptor].y == -1)
-            acceptorExclusionVector[acceptor].y = donor;
-        else if (acceptorExclusionVector[acceptor].z == -1)
-            acceptorExclusionVector[acceptor].z = donor;
-        else if (acceptorExclusionVector[acceptor].w == -1)
-            acceptorExclusionVector[acceptor].w = donor;
-        else
-            throw OpenMMException("CustomHbondForce: this platform does not support more than four exclusions per acceptor");
-    }
-    donorExclusions.initialize<mm_int4>(cc, numDonors, "customHbondDonorExclusions");
-    acceptorExclusions.initialize<mm_int4>(cc, numAcceptors, "customHbondAcceptorExclusions");
-    donorExclusions.upload(donorExclusionVector);
-    acceptorExclusions.upload(acceptorExclusionVector);
-
-    // Record the tabulated functions.
-
-    map<string, Lepton::CustomFunction*> functions;
-    vector<pair<string, string> > functionDefinitions;
-    vector<const TabulatedFunction*> functionList;
-    stringstream tableArgs;
-    tabulatedFunctionArrays.resize(force.getNumTabulatedFunctions());
-    for (int i = 0; i < force.getNumTabulatedFunctions(); i++) {
-        functionList.push_back(&force.getTabulatedFunction(i));
-        string name = force.getTabulatedFunctionName(i);
-        tabulatedFunctionUpdateCount[name] = force.getTabulatedFunction(i).getUpdateCount();
-        string arrayName = "table"+cc.intToString(i);
-        functionDefinitions.push_back(make_pair(name, arrayName));
-        functions[name] = cc.getExpressionUtilities().getFunctionPlaceholder(force.getTabulatedFunction(i));
-        int width;
-        vector<float> f = cc.getExpressionUtilities().computeFunctionCoefficients(force.getTabulatedFunction(i), width);
-        tabulatedFunctionArrays[i].initialize<float>(cc, f.size(), "TabulatedFunction");
-        tabulatedFunctionArrays[i].upload(f);
-        tableArgs << ", GLOBAL const float";
-        if (width > 1)
-            tableArgs << width;
-        tableArgs << "* RESTRICT " << arrayName;
-    }
-
-    // Record information about parameters.
-
-    globalParamNames.resize(force.getNumGlobalParameters());
-    globalParamValues.resize(force.getNumGlobalParameters());
-    for (int i = 0; i < force.getNumGlobalParameters(); i++) {
-        globalParamNames[i] = force.getGlobalParameterName(i);
-        globalParamValues[i] = (float) force.getGlobalParameterDefaultValue(i);
-    }
-    if (globals.isInitialized())
-        globals.upload(globalParamValues);
-    map<string, string> variables;
-    for (int i = 0; i < force.getNumPerDonorParameters(); i++) {
-        const string& name = force.getPerDonorParameterName(i);
-        variables[name] = "donorParams"+donorParams->getParameterSuffix(i);
-    }
-    for (int i = 0; i < force.getNumPerAcceptorParameters(); i++) {
-        const string& name = force.getPerAcceptorParameterName(i);
-        variables[name] = "acceptorParams"+acceptorParams->getParameterSuffix(i);
-    }
-    for (int i = 0; i < force.getNumGlobalParameters(); i++) {
-        const string& name = force.getGlobalParameterName(i);
-        variables[name] = "globals["+cc.intToString(i)+"]";
-    }
-
-    // Now to generate the kernel.  First, it needs to calculate all distances, angles,
-    // and dihedrals the expression depends on.
-
-    map<string, vector<int> > distances;
-    map<string, vector<int> > angles;
-    map<string, vector<int> > dihedrals;
-    Lepton::ParsedExpression energyExpression = CustomHbondForceImpl::prepareExpression(force, functions, distances, angles, dihedrals);
-    map<string, Lepton::ParsedExpression> forceExpressions;
-    set<string> computedDeltas;
-    computedDeltas.insert("D1A1");
-    string atomNames[] = {"A1", "A2", "A3", "D1", "D2", "D3"};
-    string atomNamesLower[] = {"a1", "a2", "a3", "d1", "d2", "d3"};
-    stringstream compute, extraArgs;
-    int index = 0;
-    for (auto& distance : distances) {
-        const vector<int>& atoms = distance.second;
-        string deltaName = atomNames[atoms[0]]+atomNames[atoms[1]];
-        if (computedDeltas.count(deltaName) == 0) {
-            compute << "real4 delta"+deltaName+" = delta("+atomNamesLower[atoms[0]]+", "+atomNamesLower[atoms[1]]+", periodicBoxSize, invPeriodicBoxSize, periodicBoxVecX, periodicBoxVecY, periodicBoxVecZ);\n";
-            computedDeltas.insert(deltaName);
-        }
-        compute << "real r_"+deltaName+" = SQRT(delta"+deltaName+".w);\n";
-        variables[distance.first] = "r_"+deltaName;
-        forceExpressions["real dEdDistance"+cc.intToString(index)+" = "] = energyExpression.differentiate(distance.first).optimize();
-        index++;
-    }
-    index = 0;
-    for (auto& angle : angles) {
-        const vector<int>& atoms = angle.second;
-        string deltaName1 = atomNames[atoms[1]]+atomNames[atoms[0]];
-        string deltaName2 = atomNames[atoms[1]]+atomNames[atoms[2]];
-        string angleName = "angle_"+atomNames[atoms[0]]+atomNames[atoms[1]]+atomNames[atoms[2]];
-        if (computedDeltas.count(deltaName1) == 0) {
-            compute << "real4 delta"+deltaName1+" = delta("+atomNamesLower[atoms[1]]+", "+atomNamesLower[atoms[0]]+", periodicBoxSize, invPeriodicBoxSize, periodicBoxVecX, periodicBoxVecY, periodicBoxVecZ);\n";
-            computedDeltas.insert(deltaName1);
-        }
-        if (computedDeltas.count(deltaName2) == 0) {
-            compute << "real4 delta"+deltaName2+" = delta("+atomNamesLower[atoms[1]]+", "+atomNamesLower[atoms[2]]+", periodicBoxSize, invPeriodicBoxSize, periodicBoxVecX, periodicBoxVecY, periodicBoxVecZ);\n";
-            computedDeltas.insert(deltaName2);
-        }
-        compute << "real "+angleName+" = computeAngle(delta"+deltaName1+", delta"+deltaName2+");\n";
-        variables[angle.first] = angleName;
-        forceExpressions["real dEdAngle"+cc.intToString(index)+" = "] = energyExpression.differentiate(angle.first).optimize();
-        index++;
-    }
-    index = 0;
-    for (auto& dihedral : dihedrals) {
-        const vector<int>& atoms = dihedral.second;
-        string deltaName1 = atomNames[atoms[0]]+atomNames[atoms[1]];
-        string deltaName2 = atomNames[atoms[2]]+atomNames[atoms[1]];
-        string deltaName3 = atomNames[atoms[2]]+atomNames[atoms[3]];
-        string crossName1 = "cross_"+deltaName1+"_"+deltaName2;
-        string crossName2 = "cross_"+deltaName2+"_"+deltaName3;
-        string dihedralName = "dihedral_"+atomNames[atoms[0]]+atomNames[atoms[1]]+atomNames[atoms[2]]+atomNames[atoms[3]];
-        if (computedDeltas.count(deltaName1) == 0) {
-            compute << "real4 delta"+deltaName1+" = delta("+atomNamesLower[atoms[0]]+", "+atomNamesLower[atoms[1]]+", periodicBoxSize, invPeriodicBoxSize, periodicBoxVecX, periodicBoxVecY, periodicBoxVecZ);\n";
-            computedDeltas.insert(deltaName1);
-        }
-        if (computedDeltas.count(deltaName2) == 0) {
-            compute << "real4 delta"+deltaName2+" = delta("+atomNamesLower[atoms[2]]+", "+atomNamesLower[atoms[1]]+", periodicBoxSize, invPeriodicBoxSize, periodicBoxVecX, periodicBoxVecY, periodicBoxVecZ);\n";
-            computedDeltas.insert(deltaName2);
-        }
-        if (computedDeltas.count(deltaName3) == 0) {
-            compute << "real4 delta"+deltaName3+" = delta("+atomNamesLower[atoms[2]]+", "+atomNamesLower[atoms[3]]+", periodicBoxSize, invPeriodicBoxSize, periodicBoxVecX, periodicBoxVecY, periodicBoxVecZ);\n";
-            computedDeltas.insert(deltaName3);
-        }
-        compute << "real4 "+crossName1+" = computeCross(delta"+deltaName1+", delta"+deltaName2+");\n";
-        compute << "real4 "+crossName2+" = computeCross(delta"+deltaName2+", delta"+deltaName3+");\n";
-        compute << "real "+dihedralName+" = computeAngle("+crossName1+", "+crossName2+");\n";
-        compute << dihedralName+" *= (delta"+deltaName1+".x*"+crossName2+".x + delta"+deltaName1+".y*"+crossName2+".y + delta"+deltaName1+".z*"+crossName2+".z < 0 ? -1 : 1);\n";
-        variables[dihedral.first] = dihedralName;
-        forceExpressions["real dEdDihedral"+cc.intToString(index)+" = "] = energyExpression.differentiate(dihedral.first).optimize();
-        index++;
-    }
-
-    // Next it needs to load parameters from global memory.
-
-    if (force.getNumGlobalParameters() > 0)
-        extraArgs << ", GLOBAL const float* RESTRICT globals";
-    for (int i = 0; i < (int) donorParams->getParameterInfos().size(); i++) {
-        ComputeParameterInfo& parameter = donorParams->getParameterInfos()[i];
-        extraArgs << ", GLOBAL const "+parameter.getType()+"* RESTRICT donor"+parameter.getName();
-        compute << parameter.getType()+" donorParams"+cc.intToString(i+1)+" = donor"+parameter.getName()+"[donorIndex];\n";
-    }
-    for (int i = 0; i < (int) acceptorParams->getParameterInfos().size(); i++) {
-        ComputeParameterInfo& parameter = acceptorParams->getParameterInfos()[i];
-        extraArgs << ", GLOBAL const "+parameter.getType()+"* RESTRICT acceptor"+parameter.getName();
-        compute << parameter.getType()+" acceptorParams"+cc.intToString(i+1)+" = acceptor"+parameter.getName()+"[acceptorIndex];\n";
-    }
-
-    // Now evaluate the expressions.
-
-    forceExpressions["energy += "] = energyExpression;
-    compute << cc.getExpressionUtilities().createExpressions(forceExpressions, variables, functionList, functionDefinitions, "temp");
-
-    // Finally, apply forces to atoms.
-
-    index = 0;
-    for (auto& distance : distances) {
-        const vector<int>& atoms = distance.second;
-        string deltaName = atomNames[atoms[0]]+atomNames[atoms[1]];
-        string value = "(dEdDistance"+cc.intToString(index)+"/r_"+deltaName+")*delta"+deltaName;
-        applyDonorAndAcceptorForces(compute, atoms[0], "-"+value);
-        applyDonorAndAcceptorForces(compute, atoms[1], value);
-        index++;
-    }
-    index = 0;
-    for (auto& angle : angles) {
-        const vector<int>& atoms = angle.second;
-        string deltaName1 = atomNames[atoms[1]]+atomNames[atoms[0]];
-        string deltaName2 = atomNames[atoms[1]]+atomNames[atoms[2]];
-        compute << "{\n";
-        compute << "real3 crossProd = trimTo3(cross(delta"+deltaName2+", delta"+deltaName1+"));\n";
-        compute << "real lengthCross = max(SQRT(dot(crossProd,crossProd)), (real) 1e-6f);\n";
-        compute << "real3 deltaCross0 = -cross(trimTo3(delta"+deltaName1+"), crossProd)*dEdAngle"+cc.intToString(index)+"/(delta"+deltaName1+".w*lengthCross);\n";
-        compute << "real3 deltaCross2 = cross(trimTo3(delta"+deltaName2+"), crossProd)*dEdAngle"+cc.intToString(index)+"/(delta"+deltaName2+".w*lengthCross);\n";
-        compute << "real3 deltaCross1 = -(deltaCross0+deltaCross2);\n";
-        applyDonorAndAcceptorForces(compute, atoms[0], "deltaCross0", false);
-        applyDonorAndAcceptorForces(compute, atoms[1], "deltaCross1", false);
-        applyDonorAndAcceptorForces(compute, atoms[2], "deltaCross2", false);
-        compute << "}\n";
-        index++;
-    }
-    index = 0;
-    for (auto& dihedral : dihedrals) {
-        const vector<int>& atoms = dihedral.second;
-        string deltaName1 = atomNames[atoms[0]]+atomNames[atoms[1]];
-        string deltaName2 = atomNames[atoms[2]]+atomNames[atoms[1]];
-        string deltaName3 = atomNames[atoms[2]]+atomNames[atoms[3]];
-        string crossName1 = "cross_"+deltaName1+"_"+deltaName2;
-        string crossName2 = "cross_"+deltaName2+"_"+deltaName3;
-        compute << "{\n";
-        compute << "real r = SQRT(delta"+deltaName2+".w);\n";
-        compute << "real4 ff;\n";
-        compute << "ff.x = (-dEdDihedral"+cc.intToString(index)+"*r)/"+crossName1+".w;\n";
-        compute << "ff.y = (delta"+deltaName1+".x*delta"+deltaName2+".x + delta"+deltaName1+".y*delta"+deltaName2+".y + delta"+deltaName1+".z*delta"+deltaName2+".z)/delta"+deltaName2+".w;\n";
-        compute << "ff.z = (delta"+deltaName3+".x*delta"+deltaName2+".x + delta"+deltaName3+".y*delta"+deltaName2+".y + delta"+deltaName3+".z*delta"+deltaName2+".z)/delta"+deltaName2+".w;\n";
-        compute << "ff.w = (dEdDihedral"+cc.intToString(index)+"*r)/"+crossName2+".w;\n";
-        compute << "real4 internalF0 = ff.x*"+crossName1+";\n";
-        compute << "real4 internalF3 = ff.w*"+crossName2+";\n";
-        compute << "real4 s = ff.y*internalF0 - ff.z*internalF3;\n";
-        applyDonorAndAcceptorForces(compute, atoms[0], "internalF0");
-        applyDonorAndAcceptorForces(compute, atoms[1], "s-internalF0");
-        applyDonorAndAcceptorForces(compute, atoms[2], "-s-internalF3");
-        applyDonorAndAcceptorForces(compute, atoms[3], "internalF3");
-        compute << "}\n";
-        index++;
-    }
-
-    // Generate the kernels.
-
-    map<string, string> replacements;
-    replacements["COMPUTE_FORCE"] = compute.str();
-    replacements["PARAMETER_ARGUMENTS"] = extraArgs.str()+tableArgs.str();
-    map<string, string> defines;
-    defines["PADDED_NUM_ATOMS"] = cc.intToString(cc.getPaddedNumAtoms());
-    defines["NUM_DONORS"] = cc.intToString(numDonors);
-    defines["NUM_ACCEPTORS"] = cc.intToString(numAcceptors);
-    defines["NUM_DONOR_BLOCKS"] = cc.intToString(numDonorBlocks);
-    defines["NUM_ACCEPTOR_BLOCKS"] = cc.intToString(numAcceptorBlocks);
-    defines["M_PI"] = cc.doubleToString(M_PI);
-    defines["THREAD_BLOCK_SIZE"] = "128";
-    if (force.getNonbondedMethod() != CustomHbondForce::NoCutoff) {
-        defines["USE_CUTOFF"] = "1";
-        defines["CUTOFF_SQUARED"] = cc.doubleToString(force.getCutoffDistance()*force.getCutoffDistance());
-    }
-    if (force.getNonbondedMethod() != CustomHbondForce::NoCutoff && force.getNonbondedMethod() != CustomHbondForce::CutoffNonPeriodic)
-        defines["USE_PERIODIC"] = "1";
-    if (force.getNumExclusions() > 0)
-        defines["USE_EXCLUSIONS"] = "1";
-    if (useBoundingBoxes)
-        defines["USE_BOUNDING_BOXES"] = "1";
-    ComputeProgram program = cc.compileProgram(cc.replaceStrings(CommonKernelSources::customHbondForce, replacements), defines);
-    blockBoundsKernel = program->createKernel("findBlockBounds");
-    forceKernel = program->createKernel("computeHbondForces");
-}
-
-double CommonCalcCustomHbondForceKernel::execute(ContextImpl& context, bool includeForces, bool includeEnergy) {
-    if (numDonors == 0 || numAcceptors == 0)
-        return 0.0;
-    ContextSelector selector(cc);
-    if (globals.isInitialized()) {
-        bool changed = false;
-        for (int i = 0; i < (int) globalParamNames.size(); i++) {
-            float value = (float) context.getParameter(globalParamNames[i]);
-            if (value != globalParamValues[i])
-                changed = true;
-            globalParamValues[i] = value;
-        }
-        if (changed)
-            globals.upload(globalParamValues);
-    }
-    if (!hasInitializedKernel) {
-        hasInitializedKernel = true;
-        if (useBoundingBoxes) {
-            blockBoundsKernel->addArg(donors);
-            blockBoundsKernel->addArg(acceptors);
-            for (int i = 0; i < 5; i++)
-                blockBoundsKernel->addArg(); // Periodic box size arguments are set when the kernel is executed.
-            blockBoundsKernel->addArg(cc.getPosq());
-            blockBoundsKernel->addArg(donorBlockCenter);
-            blockBoundsKernel->addArg(donorBlockSize);
-            blockBoundsKernel->addArg(acceptorBlockCenter);
-            blockBoundsKernel->addArg(acceptorBlockSize);
-        }
-        forceKernel->addArg(cc.getLongForceBuffer());
-        forceKernel->addArg(cc.getEnergyBuffer());
-        forceKernel->addArg(cc.getPosq());
-        forceKernel->addArg(donorExclusions);
-        forceKernel->addArg(donors);
-        forceKernel->addArg(acceptors);
-        for (int i = 0; i < 5; i++)
-            forceKernel->addArg(); // Periodic box size arguments are set when the kernel is executed.
-        if (useBoundingBoxes) {
-            forceKernel->addArg(donorBlockCenter);
-            forceKernel->addArg(donorBlockSize);
-            forceKernel->addArg(acceptorBlockCenter);
-            forceKernel->addArg(acceptorBlockSize);
-        }
-        if (globals.isInitialized())
-            forceKernel->addArg(globals);
-        for (auto& parameter : donorParams->getParameterInfos())
-            forceKernel->addArg(parameter.getArray());
-        for (auto& parameter : acceptorParams->getParameterInfos())
-            forceKernel->addArg(parameter.getArray());
-        for (auto& function : tabulatedFunctionArrays)
-            forceKernel->addArg(function);
-    }
-    if (useBoundingBoxes) {
-        setPeriodicBoxArgs(cc, blockBoundsKernel, 2);
-        blockBoundsKernel->execute(max(numDonors, numAcceptors));
-    }
-    setPeriodicBoxArgs(cc, forceKernel, 6);
-    int numDonorBlocks = (numDonors+31)/32;
-    int numAcceptorBlocks = (numAcceptors+31)/32;
-    forceKernel->execute(numDonorBlocks*numAcceptorBlocks*32, cc.getSIMDWidth() < 32 ? 32 : 128);
-    return 0.0;
-}
-
-void CommonCalcCustomHbondForceKernel::copyParametersToContext(ContextImpl& context, const CustomHbondForce& force) {
-    ContextSelector selector(cc);
-    int numContexts = cc.getNumContexts();
-    int startIndex = cc.getContextIndex()*force.getNumDonors()/numContexts;
-    int endIndex = (cc.getContextIndex()+1)*force.getNumDonors()/numContexts;
-    if (numDonors != endIndex-startIndex)
-        throw OpenMMException("updateParametersInContext: The number of donors has changed");
-    if (numAcceptors != force.getNumAcceptors())
-        throw OpenMMException("updateParametersInContext: The number of acceptors has changed");
-
-    // Record the per-donor parameters.
-
-    if (numDonors > 0) {
-        vector<vector<float> > donorParamVector(numDonors);
-        vector<double> parameters;
-        for (int i = 0; i < numDonors; i++) {
-            int d1, d2, d3;
-            force.getDonorParameters(startIndex+i, d1, d2, d3, parameters);
-            donorParamVector[i].resize(parameters.size());
-            for (int j = 0; j < (int) parameters.size(); j++)
-                donorParamVector[i][j] = (float) parameters[j];
-        }
-        donorParams->setParameterValues(donorParamVector);
-    }
-
-    // Record the per-acceptor parameters.
-
-    if (numAcceptors > 0) {
-        vector<vector<float> > acceptorParamVector(numAcceptors);
-        vector<double> parameters;
-        for (int i = 0; i < numAcceptors; i++) {
-            int a1, a2, a3;
-            force.getAcceptorParameters(i, a1, a2, a3, parameters);
-            acceptorParamVector[i].resize(parameters.size());
-            for (int j = 0; j < (int) parameters.size(); j++)
-                acceptorParamVector[i][j] = (float) parameters[j];
-        }
-        acceptorParams->setParameterValues(acceptorParamVector);
-    }
-
-    // See if any tabulated functions have changed.
-
-    for (int i = 0; i < force.getNumTabulatedFunctions(); i++) {
-        string name = force.getTabulatedFunctionName(i);
-        if (force.getTabulatedFunction(i).getUpdateCount() != tabulatedFunctionUpdateCount[name]) {
-            tabulatedFunctionUpdateCount[name] = force.getTabulatedFunction(i).getUpdateCount();
-            int width;
-            vector<float> f = cc.getExpressionUtilities().computeFunctionCoefficients(force.getTabulatedFunction(i), width);
-            tabulatedFunctionArrays[i].upload(f);
-        }
-    }
-
-    // Mark that the current reordering may be invalid.
-
-    cc.invalidateMolecules(info, false, true);
-}
-
-class CommonCalcCustomManyParticleForceKernel::ForceInfo : public ComputeForceInfo {
-public:
-    ForceInfo(const CustomManyParticleForce& force) : force(force) {
-    }
-    bool areParticlesIdentical(int particle1, int particle2) {
-        thread_local static vector<double> params1, params2;
-        int type1, type2;
-        force.getParticleParameters(particle1, params1, type1);
-        force.getParticleParameters(particle2, params2, type2);
-        if (type1 != type2)
-            return false;
-        for (int i = 0; i < (int) params1.size(); i++)
-            if (params1[i] != params2[i])
-                return false;
-        return true;
-    }
-    int getNumParticleGroups() {
-        return force.getNumExclusions();
-    }
-    void getParticlesInGroup(int index, vector<int>& particles) {
-        int particle1, particle2;
-        force.getExclusionParticles(index, particle1, particle2);
-        particles.resize(2);
-        particles[0] = particle1;
-        particles[1] = particle2;
-    }
-    bool areGroupsIdentical(int group1, int group2) {
-        return true;
-    }
-private:
-    const CustomManyParticleForce& force;
-};
-
-CommonCalcCustomManyParticleForceKernel::~CommonCalcCustomManyParticleForceKernel() {
-    ContextSelector selector(cc);
-    if (params != NULL)
-        delete params;
-}
-
-void CommonCalcCustomManyParticleForceKernel::initialize(const System& system, const CustomManyParticleForce& force) {
-    ContextSelector selector(cc);
-    int numParticles = force.getNumParticles();
-    int particlesPerSet = force.getNumParticlesPerSet();
-    bool centralParticleMode = (force.getPermutationMode() == CustomManyParticleForce::UniqueCentralParticle);
-    nonbondedMethod = CalcCustomManyParticleForceKernel::NonbondedMethod(force.getNonbondedMethod());
-    forceWorkgroupSize = 128;
-    findNeighborsWorkgroupSize = (cc.getSIMDWidth() >= 32 ? 128 : 32);
-    
-    // Record parameter values.
-    
-    params = new ComputeParameterSet(cc, force.getNumPerParticleParameters(), numParticles, "customManyParticleParameters");
-    vector<vector<float> > paramVector(numParticles);
-    for (int i = 0; i < numParticles; i++) {
-        vector<double> parameters;
-        int type;
-        force.getParticleParameters(i, parameters, type);
-        paramVector[i].resize(parameters.size());
-        for (int j = 0; j < (int) parameters.size(); j++)
-            paramVector[i][j] = (float) parameters[j];
-    }
-    params->setParameterValues(paramVector);
-    info = new ForceInfo(force);
-    cc.addForce(info);
-
-    // Record the tabulated functions.
-
-    map<string, Lepton::CustomFunction*> functions;
-    vector<pair<string, string> > functionDefinitions;
-    vector<const TabulatedFunction*> functionList;
-    stringstream tableArgs;
-    tabulatedFunctionArrays.resize(force.getNumTabulatedFunctions());
-    for (int i = 0; i < force.getNumTabulatedFunctions(); i++) {
-        functionList.push_back(&force.getTabulatedFunction(i));
-        string name = force.getTabulatedFunctionName(i);
-        tabulatedFunctionUpdateCount[name] = force.getTabulatedFunction(i).getUpdateCount();
-        string arrayName = "table"+cc.intToString(i);
-        functionDefinitions.push_back(make_pair(name, arrayName));
-        functions[name] = cc.getExpressionUtilities().getFunctionPlaceholder(force.getTabulatedFunction(i));
-        int width;
-        vector<float> f = cc.getExpressionUtilities().computeFunctionCoefficients(force.getTabulatedFunction(i), width);
-        tabulatedFunctionArrays[i].initialize<float>(cc, f.size(), "TabulatedFunction");
-        tabulatedFunctionArrays[i].upload(f);
-        tableArgs << ", GLOBAL const float";
-        if (width > 1)
-            tableArgs << width;
-        tableArgs << "* RESTRICT " << arrayName;
-    }
-    
-    // Record information about parameters.
-
-    globalParamNames.resize(force.getNumGlobalParameters());
-    globalParamValues.resize(force.getNumGlobalParameters());
-    for (int i = 0; i < force.getNumGlobalParameters(); i++) {
-        globalParamNames[i] = force.getGlobalParameterName(i);
-        globalParamValues[i] = (float) force.getGlobalParameterDefaultValue(i);
-    }
-    vector<pair<ExpressionTreeNode, string> > variables;
-    for (int i = 0; i < particlesPerSet; i++) {
-        string index = cc.intToString(i+1);
-        variables.push_back(makeVariable("x"+index, "pos"+index+".x"));
-        variables.push_back(makeVariable("y"+index, "pos"+index+".y"));
-        variables.push_back(makeVariable("z"+index, "pos"+index+".z"));
-    }
-    for (int i = 0; i < force.getNumPerParticleParameters(); i++) {
-        const string& name = force.getPerParticleParameterName(i);
-        for (int j = 0; j < particlesPerSet; j++) {
-            string index = cc.intToString(j+1);
-            variables.push_back(makeVariable(name+index, "((real) params"+params->getParameterSuffix(i, index)+")"));
-        }
-    }
-    if (force.getNumGlobalParameters() > 0) {
-        globals.initialize<float>(cc, force.getNumGlobalParameters(), "customManyParticleGlobals");
-        globals.upload(globalParamValues);
-        for (int i = 0; i < force.getNumGlobalParameters(); i++) {
-            const string& name = force.getGlobalParameterName(i);
-            string value = "globals["+cc.intToString(i)+"]";
-            variables.push_back(makeVariable(name, value));
-        }
-    }
-    
-    // Build data structures for type filters.
-    
-    vector<int> particleTypesVec;
-    vector<int> orderIndexVec;
-    vector<std::vector<int> > particleOrderVec;
-    int numTypes;
-    CustomManyParticleForceImpl::buildFilterArrays(force, numTypes, particleTypesVec, orderIndexVec, particleOrderVec);
-    bool hasTypeFilters = (particleOrderVec.size() > 1);
-    if (hasTypeFilters) {
-        particleTypes.initialize<int>(cc, particleTypesVec.size(), "customManyParticleTypes");
-        orderIndex.initialize<int>(cc, orderIndexVec.size(), "customManyParticleOrderIndex");
-        particleOrder.initialize<int>(cc, particleOrderVec.size()*particlesPerSet, "customManyParticleOrder");
-        particleTypes.upload(particleTypesVec);
-        orderIndex.upload(orderIndexVec);
-        vector<int> flattenedOrder(particleOrder.getSize());
-        for (int i = 0; i < (int) particleOrderVec.size(); i++)
-            for (int j = 0; j < particlesPerSet; j++)
-                flattenedOrder[i*particlesPerSet+j] = particleOrderVec[i][j];
-        particleOrder.upload(flattenedOrder);
-    }
-    
-    // Build data structures for exclusions.
-    
-    if (force.getNumExclusions() > 0) {
-        vector<vector<int> > particleExclusions(numParticles);
-        for (int i = 0; i < force.getNumExclusions(); i++) {
-            int p1, p2;
-            force.getExclusionParticles(i, p1, p2);
-            particleExclusions[p1].push_back(p2);
-            particleExclusions[p2].push_back(p1);
-        }
-        vector<int> exclusionsVec;
-        vector<int> exclusionStartIndexVec(numParticles+1);
-        exclusionStartIndexVec[0] = 0;
-        for (int i = 0; i < numParticles; i++) {
-            sort(particleExclusions[i].begin(), particleExclusions[i].end());
-            exclusionsVec.insert(exclusionsVec.end(), particleExclusions[i].begin(), particleExclusions[i].end());
-            exclusionStartIndexVec[i+1] = exclusionsVec.size();
-        }
-        exclusions.initialize<int>(cc, exclusionsVec.size(), "customManyParticleExclusions");
-        exclusionStartIndex.initialize<int>(cc, exclusionStartIndexVec.size(), "customManyParticleExclusionStart");
-        exclusions.upload(exclusionsVec);
-        exclusionStartIndex.upload(exclusionStartIndexVec);
-    }
-    
-    // Build data structures for the neighbor list.
-    
-    int numAtomBlocks = cc.getPaddedNumAtoms()/32;
-    if (nonbondedMethod != NoCutoff) {
-        int elementSize = (cc.getUseDoublePrecision() ? sizeof(double) : sizeof(float));
-        blockCenter.initialize(cc, numAtomBlocks, 4*elementSize, "blockCenter");
-        blockBoundingBox.initialize(cc, numAtomBlocks, 4*elementSize, "blockBoundingBox");
-        numNeighborPairs.initialize<int>(cc, 1, "customManyParticleNumNeighborPairs");
-        neighborStartIndex.initialize<int>(cc, numParticles+1, "customManyParticleNeighborStartIndex");
-        numNeighborsForAtom.initialize<int>(cc, numParticles, "customManyParticleNumNeighborsForAtom");
-
-        // Select a size for the array that holds the neighbor list.  We have to make a fairly
-        // arbitrary guess, but if this turns out to be too small we'll increase it later.
-
-        maxNeighborPairs = 150*numParticles;
-        neighborPairs.initialize<mm_int2>(cc, maxNeighborPairs, "customManyParticleNeighborPairs");
-        neighbors.initialize<int>(cc, maxNeighborPairs, "customManyParticleNeighbors");
-    }
-
-    // Generate the kernel.
-
-    Lepton::ParsedExpression energyExpression = CustomManyParticleForceImpl::prepareExpression(force, functions);
-    map<string, Lepton::ParsedExpression> forceExpressions;
-    stringstream compute;
-    for (int i = 0; i < (int) params->getParameterInfos().size(); i++) {
-        ComputeParameterInfo& parameter = params->getParameterInfos()[i];
-        compute<<parameter.getType()<<" params"<<(i+1)<<" = global_params"<<(i+1)<<"[index];\n";
-    }
-    forceExpressions["energy += "] = energyExpression;
-    vector<string> forceNames;
-    for (int i = 0; i < particlesPerSet; i++) {
-        string istr = cc.intToString(i+1);
-        string forceName = "force"+istr;
-        forceNames.push_back(forceName);
-        compute<<"real3 "<<forceName<<" = make_real3(0);\n";
-        Lepton::ParsedExpression forceExpressionX = energyExpression.differentiate("x"+istr).optimize();
-        Lepton::ParsedExpression forceExpressionY = energyExpression.differentiate("y"+istr).optimize();
-        Lepton::ParsedExpression forceExpressionZ = energyExpression.differentiate("z"+istr).optimize();
-        if (!isZeroExpression(forceExpressionX))
-            forceExpressions[forceName+".x -= "] = forceExpressionX;
-        if (!isZeroExpression(forceExpressionY))
-            forceExpressions[forceName+".y -= "] = forceExpressionY;
-        if (!isZeroExpression(forceExpressionZ))
-            forceExpressions[forceName+".z -= "] = forceExpressionZ;
-    }
-    compute << cc.getExpressionUtilities().createExpressions(forceExpressions, variables, functionList, functionDefinitions, "temp", "real", force.usesPeriodicBoundaryConditions());
-    
-    // Store forces to global memory.
-    
-    for (int i = 0; i < particlesPerSet; i++)
-        compute<<"storeForce(atom"<<(i+1)<<", "<<forceNames[i]<<", forceBuffers);\n";
-    
-    // Create other replacements that depend on the number of particles per set.
-    
-    stringstream numCombinations, atomsForCombination, isValidCombination, permute, loadData, verifyCutoff, verifyExclusions;
-    if (hasTypeFilters) {
-        permute<<"int particleSet[] = {";
-        for (int i = 0; i < particlesPerSet; i++) {
-            permute<<"p"<<(i+1);
-            if (i < particlesPerSet-1)
-                permute<<", ";
-        }
-        permute<<"};\n";
-    }
-    for (int i = 0; i < particlesPerSet; i++) {
-        if (hasTypeFilters)
-            permute<<"int atom"<<(i+1)<<" = particleSet[particleOrder["<<particlesPerSet<<"*order+"<<i<<"]];\n";
-        else
-            permute<<"int atom"<<(i+1)<<" = p"<<(i+1)<<";\n";
-        loadData<<"real3 pos"<<(i+1)<<" = trimTo3(posq[atom"<<(i+1)<<"]);\n";
-        for (int j = 0; j < (int) params->getParameterInfos().size(); j++)
-            loadData<<params->getParameterInfos()[j].getType()<<" params"<<(j+1)<<(i+1)<<" = global_params"<<(j+1)<<"[atom"<<(i+1)<<"];\n";
-    }
-    if (centralParticleMode) {
-        for (int i = 1; i < particlesPerSet; i++) {
-            if (i > 1)
-                isValidCombination<<" && p"<<(i+1)<<">p"<<i<<" && ";
-            isValidCombination<<"p"<<(i+1)<<"!=p1";
-        }
-    }
-    else {
-        for (int i = 2; i < particlesPerSet; i++) {
-            if (i > 2)
-                isValidCombination<<" && ";
-            isValidCombination<<"a"<<(i+1)<<">a"<<i;
-        }
-    }
-    atomsForCombination<<"int tempIndex = index;\n";
-    for (int i = 1; i < particlesPerSet; i++) {
-        if (i > 1)
-            numCombinations<<"*";
-        numCombinations<<"numNeighbors";
-        if (centralParticleMode)
-            atomsForCombination<<"int a"<<(i+1)<<" = tempIndex%numNeighbors;\n";
-        else
-            atomsForCombination<<"int a"<<(i+1)<<" = 1+tempIndex%numNeighbors;\n";
-        if (i < particlesPerSet-1)
-            atomsForCombination<<"tempIndex /= numNeighbors;\n";
-    }
-    if (particlesPerSet > 2) {
-        if (centralParticleMode)
-            atomsForCombination<<"a2 = (a3%2 == 0 ? a2 : numNeighbors-a2-1);\n";
-        else
-            atomsForCombination<<"a2 = (a3%2 == 0 ? a2 : numNeighbors-a2+1);\n";
-    }
-    for (int i = 1; i < particlesPerSet; i++) {
-        if (nonbondedMethod == NoCutoff) {
-            if (centralParticleMode)
-                atomsForCombination<<"int p"<<(i+1)<<" = a"<<(i+1)<<";\n";
-            else
-                atomsForCombination<<"int p"<<(i+1)<<" = p1+a"<<(i+1)<<";\n";
-        }
-        else {
-            if (centralParticleMode)
-                atomsForCombination<<"int p"<<(i+1)<<" = neighbors[firstNeighbor+a"<<(i+1)<<"];\n";
-            else
-                atomsForCombination<<"int p"<<(i+1)<<" = neighbors[firstNeighbor-1+a"<<(i+1)<<"];\n";
-        }
-    }
-    if (nonbondedMethod != NoCutoff) {
-        for (int i = 1; i < particlesPerSet; i++)
-            verifyCutoff<<"real3 pos"<<(i+1)<<" = trimTo3(posq[p"<<(i+1)<<"]);\n";
-        if (!centralParticleMode) {
-            for (int i = 1; i < particlesPerSet; i++) {
-                for (int j = i+1; j < particlesPerSet; j++)
-                    verifyCutoff<<"includeInteraction &= (delta(pos"<<(i+1)<<", pos"<<(j+1)<<", periodicBoxSize, invPeriodicBoxSize, periodicBoxVecX, periodicBoxVecY, periodicBoxVecZ).w < CUTOFF_SQUARED);\n";
-            }
-        }
-    }
-    if (force.getNumExclusions() > 0) {
-        int startCheckFrom = (nonbondedMethod == NoCutoff ? 0 : 1);
-        for (int i = startCheckFrom; i < particlesPerSet; i++)
-            for (int j = i+1; j < particlesPerSet; j++)
-                verifyExclusions<<"includeInteraction &= !isInteractionExcluded(p"<<(i+1)<<", p"<<(j+1)<<", exclusions, exclusionStartIndex);\n";
-    }
-    string computeTypeIndex = "particleTypes[p"+cc.intToString(particlesPerSet)+"]";
-    for (int i = particlesPerSet-2; i >= 0; i--)
-        computeTypeIndex = "particleTypes[p"+cc.intToString(i+1)+"]+"+cc.intToString(numTypes)+"*("+computeTypeIndex+")";
-    
-    // Create replacements for extra arguments.
-    
-    stringstream extraArgs;
-    if (force.getNumGlobalParameters() > 0)
-        extraArgs << ", GLOBAL const float* globals";
-    for (int i = 0; i < (int) params->getParameterInfos().size(); i++) {
-        ComputeParameterInfo& parameter = params->getParameterInfos()[i];
-        extraArgs<<", GLOBAL const "<<parameter.getType()<<"* RESTRICT global_params"<<(i+1);
-    }
-
-    // Create the kernels.
-
-    map<string, string> replacements;
-    replacements["COMPUTE_INTERACTION"] = compute.str();
-    replacements["NUM_CANDIDATE_COMBINATIONS"] = numCombinations.str();
-    replacements["FIND_ATOMS_FOR_COMBINATION_INDEX"] = atomsForCombination.str();
-    replacements["IS_VALID_COMBINATION"] = isValidCombination.str();
-    replacements["VERIFY_CUTOFF"] = verifyCutoff.str();
-    replacements["VERIFY_EXCLUSIONS"] = verifyExclusions.str();
-    replacements["PERMUTE_ATOMS"] = permute.str();
-    replacements["LOAD_PARTICLE_DATA"] = loadData.str();
-    replacements["COMPUTE_TYPE_INDEX"] = computeTypeIndex;
-    replacements["PARAMETER_ARGUMENTS"] = extraArgs.str()+tableArgs.str();
-    map<string, string> defines;
-    if (nonbondedMethod != NoCutoff)
-        defines["USE_CUTOFF"] = "1";
-    if (nonbondedMethod == CutoffPeriodic)
-        defines["USE_PERIODIC"] = "1";
-    if (centralParticleMode)
-        defines["USE_CENTRAL_PARTICLE"] = "1";
-    if (hasTypeFilters)
-        defines["USE_FILTERS"] = "1";
-    if (force.getNumExclusions() > 0)
-        defines["USE_EXCLUSIONS"] = "1";
-    defines["NUM_ATOMS"] = cc.intToString(cc.getNumAtoms());
-    defines["PADDED_NUM_ATOMS"] = cc.intToString(cc.getPaddedNumAtoms());
-    defines["M_PI"] = cc.doubleToString(M_PI);
-    defines["CUTOFF_SQUARED"] = cc.doubleToString(force.getCutoffDistance()*force.getCutoffDistance());
-    defines["TILE_SIZE"] = cc.intToString(32);
-    defines["NUM_BLOCKS"] = cc.intToString(numAtomBlocks);
-    defines["FIND_NEIGHBORS_WORKGROUP_SIZE"] = cc.intToString(findNeighborsWorkgroupSize);
-    ComputeProgram program = cc.compileProgram(cc.replaceStrings(CommonKernelSources::pointFunctions+CommonKernelSources::customManyParticle, replacements), defines);
-    forceKernel = program->createKernel("computeInteraction");
-    blockBoundsKernel = program->createKernel("findBlockBounds");
-    neighborsKernel = program->createKernel("findNeighbors");
-    startIndicesKernel = program->createKernel("computeNeighborStartIndices");
-    copyPairsKernel = program->createKernel("copyPairsToNeighborList");
-    event = cc.createEvent();
-}
-
-double CommonCalcCustomManyParticleForceKernel::execute(ContextImpl& context, bool includeForces, bool includeEnergy) {
-    ContextSelector selector(cc);
-    if (!hasInitializedKernel) {
-        hasInitializedKernel = true;
-        
-        // Set arguments for the force kernel.
-        
-        forceKernel->addArg(cc.getLongForceBuffer());
-        forceKernel->addArg(cc.getEnergyBuffer());
-        forceKernel->addArg(cc.getPosq());
-        for (int i = 0; i < 5; i++)
-            forceKernel->addArg();
-        setPeriodicBoxArgs(cc, forceKernel, 3);
-        if (nonbondedMethod != NoCutoff) {
-            forceKernel->addArg(neighbors);
-            forceKernel->addArg(neighborStartIndex);
-        }
-        if (particleTypes.isInitialized()) {
-            forceKernel->addArg(particleTypes);
-            forceKernel->addArg(orderIndex);
-            forceKernel->addArg(particleOrder);
-        }
-        if (exclusions.isInitialized()) {
-            forceKernel->addArg(exclusions);
-            forceKernel->addArg(exclusionStartIndex);
-        }
-        if (globals.isInitialized())
-            forceKernel->addArg(globals);
-        for (auto& parameter : params->getParameterInfos())
-            forceKernel->addArg(parameter.getArray());
-        for (auto& function : tabulatedFunctionArrays)
-            forceKernel->addArg(function);
-        
-        if (nonbondedMethod != NoCutoff) {
-            // Set arguments for the block bounds kernel.
-
-            for (int i = 0; i < 5; i++)
-                blockBoundsKernel->addArg(); // Periodic box information will be set just before it is executed.
-            blockBoundsKernel->addArg(cc.getPosq());
-            blockBoundsKernel->addArg(blockCenter);
-            blockBoundsKernel->addArg(blockBoundingBox);
-            blockBoundsKernel->addArg(numNeighborPairs);
-
-            // Set arguments for the neighbor list kernel.
-
-            for (int i = 0; i < 5; i++)
-                neighborsKernel->addArg(); // Periodic box information will be set just before it is executed.
-            neighborsKernel->addArg(cc.getPosq());
-            neighborsKernel->addArg(blockCenter);
-            neighborsKernel->addArg(blockBoundingBox);
-            neighborsKernel->addArg(neighborPairs);
-            neighborsKernel->addArg(numNeighborPairs);
-            neighborsKernel->addArg(numNeighborsForAtom);
-            neighborsKernel->addArg(maxNeighborPairs);
-            if (exclusions.isInitialized()) {
-                neighborsKernel->addArg(exclusions);
-                neighborsKernel->addArg(exclusionStartIndex);
-            }
-            
-            // Set arguments for the kernel to find neighbor list start indices.
-            
-            startIndicesKernel->addArg(numNeighborsForAtom);
-            startIndicesKernel->addArg(neighborStartIndex);
-            startIndicesKernel->addArg(numNeighborPairs);
-            startIndicesKernel->addArg(maxNeighborPairs);
-
-            // Set arguments for the kernel to assemble the final neighbor list.
-            
-            copyPairsKernel->addArg(neighborPairs);
-            copyPairsKernel->addArg(neighbors);
-            copyPairsKernel->addArg(numNeighborPairs);
-            copyPairsKernel->addArg(maxNeighborPairs);
-            copyPairsKernel->addArg(numNeighborsForAtom);
-            copyPairsKernel->addArg(neighborStartIndex);
-       }
-    }
-    if (globals.isInitialized()) {
-        bool changed = false;
-        for (int i = 0; i < (int) globalParamNames.size(); i++) {
-            float value = (float) context.getParameter(globalParamNames[i]);
-            if (value != globalParamValues[i])
-                changed = true;
-            globalParamValues[i] = value;
-        }
-        if (changed)
-            globals.upload(globalParamValues);
-    }
-    while (true) {
-        int* numPairs = (int*) cc.getPinnedBuffer();
-        if (nonbondedMethod != NoCutoff) {
-            setPeriodicBoxArgs(cc, forceKernel, 3);
-            setPeriodicBoxArgs(cc, blockBoundsKernel, 0);
-            setPeriodicBoxArgs(cc, neighborsKernel, 0);
-            blockBoundsKernel->execute(cc.getPaddedNumAtoms()/32);
-            neighborsKernel->execute(cc.getNumAtoms(), findNeighborsWorkgroupSize);
-
-            // We need to make sure there was enough memory for the neighbor list.  Download the
-            // information asynchronously so kernels can be running at the same time.
-
-            numNeighborPairs.download(numPairs, false);
-            event->enqueue();
-            startIndicesKernel->execute(256, 256);
-            copyPairsKernel->execute(maxNeighborPairs);
-        }
-        int maxThreads = min(cc.getNumAtoms()*forceWorkgroupSize, (int) cc.getEnergyBuffer().getSize());
-        forceKernel->execute(maxThreads, forceWorkgroupSize);
-        if (nonbondedMethod != NoCutoff) {
-            // Make sure there was enough memory for the neighbor list.
-
-            event->wait();
-            if (*numPairs > maxNeighborPairs) {
-                // Resize the arrays and run the calculation again.
-
-                maxNeighborPairs = (int) (1.1*(*numPairs));
-                neighborPairs.resize(maxNeighborPairs);
-                neighbors.resize(maxNeighborPairs);
-                neighborsKernel->setArg(11, maxNeighborPairs);
-                startIndicesKernel->setArg(3, maxNeighborPairs);
-                copyPairsKernel->setArg(3, maxNeighborPairs);
-                continue;
-            }
-        }
-        break;
-    }
-    return 0.0;
-}
-
-void CommonCalcCustomManyParticleForceKernel::copyParametersToContext(ContextImpl& context, const CustomManyParticleForce& force) {
-    ContextSelector selector(cc);
-    int numParticles = force.getNumParticles();
-    if (numParticles != cc.getNumAtoms())
-        throw OpenMMException("updateParametersInContext: The number of particles has changed");
-
-    // Record the per-particle parameters.
-
-    vector<vector<float> > paramVector(numParticles);
-    vector<double> parameters;
-    int type;
-    for (int i = 0; i < numParticles; i++) {
-        force.getParticleParameters(i, parameters, type);
-        paramVector[i].resize(parameters.size());
-        for (int j = 0; j < (int) parameters.size(); j++)
-            paramVector[i][j] = (float) parameters[j];
-    }
-    params->setParameterValues(paramVector);
-
-    // See if any tabulated functions have changed.
-
-    for (int i = 0; i < force.getNumTabulatedFunctions(); i++) {
-        string name = force.getTabulatedFunctionName(i);
-        if (force.getTabulatedFunction(i).getUpdateCount() != tabulatedFunctionUpdateCount[name]) {
-            tabulatedFunctionUpdateCount[name] = force.getTabulatedFunction(i).getUpdateCount();
-            int width;
-            vector<float> f = cc.getExpressionUtilities().computeFunctionCoefficients(force.getTabulatedFunction(i), width);
-            tabulatedFunctionArrays[i].upload(f);
-        }
-    }
-
-    // Mark that the current reordering may be invalid.
-
-    cc.invalidateMolecules(info);
 }
 
 class CommonCalcGayBerneForceKernel::ForceInfo : public ComputeForceInfo {
